[any message]

+++ src/main.rs
@@ -2,13 +2,13 @@ use std::env;
 // use serde_bencode
 
 #[allow(dead_code)]
-fn decode_bencoded_value(encoded_value: &str) -> i32 {
+fn decode_bencoded_value(encoded_value: &str) -> i64 {
 
     let string_ecode = std::string::String::from(encoded_value);
     let uncoded_value = string_ecode.trim_matches(['i','e']);
     // If encoded_value starts with a digit, it's a number
     if uncoded_value.chars().next().unwrap().is_digit(10) {
-        return uncoded_value.parse::<i32>().unwrap();
+        return uncoded_value.parse::<i64>().unwrap();
     } else {
         panic!("Unhandled encoded value: {}", encoded_value)
     }
